refactor: delay class로 변경 및 파일 경로 수정
