+++ public/js/view/news.js
@@ -1,5 +1,4 @@
-import delay from "./delay.js";
-
+import Delay from "../components/delay.js";
 
 async function getNewsData() {
   try {
@@ -29,31 +28,9 @@ async function getNewsData() {
   }
 }
 
-// function titleEventListener(titleList, newsArticle, data) {
-//   // let clickTitle;
-
-//   titleList.addEventListener("click", async function (event) {
-//     // if (clickTitle !== null) {
-//     //   clearTimeout(clickTitle);
-//     //   clickTitle = null; // delay 사용 후 clear 해줄 때 null 처리 꼭
-//     // }
-
-//     const { target } = event;
-//     if (target.tagName === "P") {
-//       const index = Array.from(this.children).indexOf(event.target);
-//       const { title, description } = data[index];
-//       newsArticle.innerHTML = `<div class="loading-text">Loading ...</div>`;
-
-//       // const delay = (ms) => new Promise((resolve) => {
-//       //     clickTitle = setTimeout(resolve, ms);
-//       // });
-
-//       await delay(3000);
-//       newsArticle.innerHTML = `<h3>${title}</h3><p>${description}</p>`;
-//     }
-//   });
-// }
 async function titleEventListener(titleList, newsArticle, data) {
+  const newsDelay = new Delay();
+
   titleList.addEventListener("click", async function (event) {
     const { target } = event;
     if (target.tagName === "P") {
@@ -61,7 +38,8 @@ async function titleEventListener(titleList, newsArticle, data) {
       const { title, description } = data[index];
       newsArticle.innerHTML = `<div class="loading-text">Loading ...</div>`;
 
-      await delay(3000);
+      await newsDelay.setDelay(3000);
+
       newsArticle.innerHTML = `<h3>${title}</h3><p>${description}</p>`;
     }
   });
