+++ public/js/view/timer.js
@@ -1,27 +1,18 @@
+import Delay from "../components/delay.js";
+
 function updateTimer(seconds) {
   const timerElement = document.querySelector(".timer");
   timerElement.textContent = `${seconds}초 뒤에 업데이트`;
 }
 
-let currentDelay = null;
-
-const delay = (ms) => {
-  if (currentDelay !== null) {
-    clearTimeout(currentDelay);
-    currentDelay = null;
-  }
-  return new Promise((resolve) => {
-    currentDelay = setTimeout(resolve, ms);
-  });
-};
-
 async function countdown() {
   let seconds = 60;
-  
+  const timerDelay = new Delay();
+
   updateTimer(seconds);
 
   while (seconds > 0) {
-    await delay(1000);
+    await timerDelay.setDelay(1000);
     seconds--;
     updateTimer(seconds);
   }
@@ -29,6 +20,5 @@ async function countdown() {
   const updateBtn = document.querySelector(".update");
   updateBtn.click();
 }
-
 
 export default countdown;
